+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+### 08/05/2017 - 1.0.4
+* Added `orderBy()` (sorting field and direction `ASC` and `DESC`)
+* Added `limit()` Limit results returned, includes Limit and Offset options.
+
 ### 08/05/2017 - 1.0.3
 * Added the `NOT LIKE` operator